Implement Time Save Columns
There's now "Possible Time Save" as a Column Start With setting
and "Time Saved or Lost" as a Column Update With setting. Tests
are still missing and the update with colors are unimplemented.

Resolves #177
